Jitter bug fixed. Next step optimise for less I/O
The jitter bug was due to 4-byte real in the DISORT run. Edited to make
all variables in F90 double, and compile F77 with 8-byte real default.
Jitter gone. Isothermal atmosphere with no opacities is now 1% below
black body curve. If we add CIA and lines we have some residual CIA
apparent in the curve too, still based at 1% below black body.

Don't understand this.

DISORT now has most output surpressed, but can't kill the header
lines. Some speed up as a result. Can't run 2stream in DISORT.

Now will optimise to take opacities into memory for speed up...

+++ atmos_ops_mod.f90
@@ -23,7 +23,7 @@ contains
     double precision :: p1,p2
     real, intent(IN) :: grav
     double precision, dimension(nlayers), intent(IN) :: press
-    real, dimension(nlayers), intent(IN):: temp
+    double precision, dimension(nlayers), intent(IN):: temp
     double precision, dimension(nlayers), intent(OUT):: dz
     
     
@@ -69,8 +69,8 @@ contains
     ! Tbars that are in our atmospheric layers.
     ! This is kept in atmos_ops for consistency a
     integer :: i
-    real, dimension(nlayers),INTENT(IN):: layertemp
-    real,INTENT(OUT):: leveltemp(0:nlayers)
+    double precision, dimension(nlayers),INTENT(IN):: layertemp
+    double precision,INTENT(OUT):: leveltemp(0:nlayers)
 
 
     leveltemp(0) = layertemp(1) - 0.5*(layertemp(2) - layertemp(1))
